Use apiGroupName in generated package for OpenShiftClient Handler/OperationsImpl
Right now all OpenShiftClient's Handler/OperationsImpl classes get
generated in one package `io.fabric8.openshift.client.handlers`/`io.fabric8.openshift.client.dsl.internal`

This creates problems if we want to provide handlers for resources with
same names but from different packages. Hence, we should use apiGroup
name in packages in order to segregate handlers/operations classes as
per their respective apiGroup
