+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Fix #3057: Removed debug calls for CustomResource during deserialization
 * Fix #3050: More enforcement of the informer lifecycle
 * Fix #3061: Removed the deltafifo from the informer logic
+* Use apiGroupName in generated package for OpenShiftClient Handler/OperationsImpl classes
 
 #### Dependency Upgrade
 * Fix #2979: Update Kubernetes Model to v1.21.0
