+++ openshift-client/src/main/java/io/fabric8/openshift/client/DefaultOpenShiftClient.java
@@ -184,31 +184,31 @@ import io.fabric8.openshift.client.dsl.OpenShiftQuotaAPIGroupDSL;
 import io.fabric8.openshift.client.dsl.ProjectOperation;
 import io.fabric8.openshift.client.dsl.ProjectRequestOperation;
 import io.fabric8.openshift.client.dsl.TemplateResource;
-import io.fabric8.openshift.client.dsl.internal.BuildConfigOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.BuildOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ClusterNetworkOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ClusterRoleBindingOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.DeploymentConfigOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.EgressNetworkPolicyOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.GroupOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ImageOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ImageStreamOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ImageStreamTagOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ImageTagOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.NetNamespaceOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.OAuthAccessTokenOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.OAuthAuthorizeTokenOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.OAuthClientOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.build.BuildConfigOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.build.BuildOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.network.ClusterNetworkOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.authorization.ClusterRoleBindingOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.apps.DeploymentConfigOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.network.EgressNetworkPolicyOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.user.GroupOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.image.ImageOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.image.ImageStreamOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.image.ImageStreamTagOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.image.ImageTagOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.network.NetNamespaceOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.oauth.OAuthAccessTokenOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.oauth.OAuthAuthorizeTokenOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.oauth.OAuthClientOperationsImpl;
 import io.fabric8.openshift.client.dsl.internal.OpenShiftSubjectAccessReviewOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ProjectOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.project.ProjectOperationsImpl;
 import io.fabric8.openshift.client.dsl.internal.ProjectRequestsOperationImpl;
-import io.fabric8.openshift.client.dsl.internal.RangeAllocationOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.RoleBindingOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.RoleOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.RouteOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.SecurityContextConstraintsOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.TemplateOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.UserOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.security.RangeAllocationOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.authorization.RoleBindingOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.authorization.RoleOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.route.RouteOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.security.SecurityContextConstraintsOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.core.TemplateOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.user.UserOperationsImpl;
 import io.fabric8.openshift.client.internal.OpenShiftClusterOperationsImpl;
 import io.fabric8.openshift.client.internal.OpenShiftNamespaceVisitFromServerGetWatchDeleteRecreateWaitApplicableListImpl;
 import io.fabric8.openshift.client.internal.OpenShiftOAuthInterceptor;
