+++ openshift-client/src/main/java/io/fabric8/openshift/client/OpenShiftOperatorAPIGroupClient.java
@@ -59,25 +59,25 @@ import io.fabric8.openshift.api.model.operator.v1.ServiceCatalogControllerManage
 import io.fabric8.openshift.api.model.operator.v1alpha1.ImageContentSourcePolicy;
 import io.fabric8.openshift.api.model.operator.v1alpha1.ImageContentSourcePolicyList;
 import io.fabric8.openshift.client.dsl.OpenShiftOperatorAPIGroupDSL;
-import io.fabric8.openshift.client.dsl.internal.AuthenticationOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.CSISnapshotControllerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ConsoleOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.DNSOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.DNSRecordOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.EtcdOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ImageContentSourcePolicyOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ImagePrunerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.IngressControllerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.KubeAPIServerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.KubeControllerManagerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.KubeSchedulerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.KubeStorageVersionMigratorOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.NetworkOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.OpenShiftAPIServerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.OpenShiftControllerManagerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ServiceCAOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ServiceCatalogAPIServerOperationsImpl;
-import io.fabric8.openshift.client.dsl.internal.ServiceCatalogControllerManagerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.AuthenticationOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.CSISnapshotControllerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.ConsoleOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.DNSOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.ingress.operator.DNSRecordOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.EtcdOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.ImageContentSourcePolicyOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.imageregistry.operator.ImagePrunerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.IngressControllerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.KubeAPIServerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.KubeControllerManagerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.KubeSchedulerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.KubeStorageVersionMigratorOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.NetworkOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.OpenShiftAPIServerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.OpenShiftControllerManagerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.ServiceCAOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.ServiceCatalogAPIServerOperationsImpl;
+import io.fabric8.openshift.client.dsl.internal.operator.ServiceCatalogControllerManagerOperationsImpl;
 import okhttp3.OkHttpClient;
 
 public class OpenShiftOperatorAPIGroupClient extends BaseClient implements OpenShiftOperatorAPIGroupDSL {
