+++ openshift-client/src/main/java/io/fabric8/openshift/client/dsl/internal/oauth/OAuthAuthorizeTokenOperationsImpl.java
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.openshift.client.dsl.internal;
+package io.fabric8.openshift.client.dsl.internal.oauth;
 
 import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.openshift.api.model.OAuthAuthorizeToken;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.base.OperationContext;
+import io.fabric8.openshift.client.dsl.internal.OpenShiftOperation;
 import okhttp3.OkHttpClient;
 import io.fabric8.openshift.api.model.OAuthAuthorizeTokenBuilder;
 import io.fabric8.openshift.api.model.OAuthAuthorizeTokenList;
